+++ vector/src/main/java/im/vector/util/ExternalApplicationsUtil.kt
@@ -17,6 +17,7 @@
 package im.vector.util
 
 import android.app.Activity
+import android.app.Fragment
 import android.content.ActivityNotFoundException
 import android.content.ContentValues
 import android.content.Context
@@ -78,17 +79,30 @@ fun openSoundRecorder(activity: Activity, requestCode: Int) {
 /**
  * Open file selection activity
  */
-fun openFileSelection(activity: Activity, requestCode: Int) {
-    val fileIntent = Intent(Intent.ACTION_GET_CONTENT)
+fun openFileSelection(activity: Activity,
+                      fragment: Fragment?,
+                      allowMultipleSelection: Boolean,
+                      requestCode: Int) {
+    val fileIntent = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+        Intent(Intent.ACTION_OPEN_DOCUMENT)
+    } else {
+        Intent(Intent.ACTION_GET_CONTENT)
+    }
 
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
-        fileIntent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true)
+        fileIntent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, allowMultipleSelection)
     }
+
+    fileIntent.addCategory(Intent.CATEGORY_OPENABLE);
 
     fileIntent.type = "*/*"
 
     try {
-        activity.startActivityForResult(fileIntent, requestCode)
+        fragment
+                ?.startActivityForResult(fileIntent, requestCode)
+                ?: run {
+                    activity.startActivityForResult(fileIntent, requestCode)
+                }
     } catch (activityNotFoundException: ActivityNotFoundException) {
         activity.toast(R.string.error_no_external_application_found)
     }